tradingbot-backend/accounthandler/src/requirements.txt
python-binance
tinydb
fastapi
python-dotenv